Add files via upload
Tweaked 'error' blink so that it wouldn't error blink if battery was charged. (Changed != 0 to >1) and set the led to turn on initially so it would be easy to identify a lack of CAN comms or bad Shield comms (which seems to happen and requires an Arduino power reset).  I'll change the code to just display status on an LCD probably someday.